+++ lib/twine/cli.rb
@@ -45,6 +45,14 @@ module Twine
           files are UTF-16 without BOM, you need to specify if it's UTF-16LE or UTF16-BE.
         DESC
       },
+      escape_all_tags: {
+        switch: ['--[no-]escape-all-tags'],
+        description: <<-DESC,
+          Always escape all HTML tags. By default the Android formatter will ONLY escape styling tags, if a 
+          string also contains placeholders. This flag enforces that styling tags are escaped regardless of
+          placeholders.
+        DESC
+      },
       file_name: {
         switch: ['-n', '--file-name FILE_NAME'],
         description: 'This flag may be used to overwrite the default file name of the format.'
@@ -111,6 +119,7 @@ module Twine
         optional_options: [
           :developer_language,
           :encoding,
+          :escape_all_tags,
           :format,
           :include,
           :languages,
@@ -133,6 +142,7 @@ module Twine
           :create_folders,
           :developer_language,
           :encoding,
+          :escape_all_tags,
           :file_name,
           :format,
           :include,
@@ -152,6 +162,7 @@ module Twine
         optional_options: [
           :developer_language,
           :encoding,
+          :escape_all_tags,
           :include,
           :quiet,
           :tags,